Got Ship Flying
The ship now flies around and properly wraps around the screen.

For some reason, if you are holding a key down, it keeps sending a
repeat keydown event. However, as soon as another key is pressed, the
repeat event for the prior never gets sent again.

So, the ship only maintains current velocity, if you rotate while
holding the ‘up’ key down, you won’t change flight direction; you just
spin on the same trajectory.

Working on implementing polling of the keyboard directly over using the
keyboard events; but polling happens way quicker than the events were
getting triggered (probably from the system’s key repeat delay or
whatever). So a timer or something is going to need to be implemented
to slow it down.